Fix more conversion errors

+++ example_10/hardware/rrbot.cpp
@@ -194,7 +194,7 @@ hardware_interface::return_type RRBotSystemWithGPIOHardware::read(
   }
 
   // random inputs analog_input1 and analog_input2
-  unsigned int seed = time(NULL) + 1;
+  unsigned int seed = static_cast<unsigned int>(time(NULL)) + 1;
   set_state(
     info_.gpios[0].name + "/" + info_.gpios[0].state_interfaces[1].name,
     static_cast<float>(rand_r(&seed)));
